+++ models/Stupid_Project/stupid03.cc
@@ -104,7 +104,7 @@ void StupidModel::addBugs(TCL_args args)
 
 void StupidModel::moveBugs()
 {
-  for (int i=0; i<bugs.size(); i++)
+  for (size_t i=0; i<bugs.size(); i++)
     bugs[i]->move();
   tstep++;
 }
@@ -140,7 +140,7 @@ void StupidModel::draw(TCL_args args)
   eco_string canvas=args;
   tclcmd c;
   c << canvas << "delete bugs\n";
-  for (int i=0; i<bugs.size(); i++)
+  for (size_t i=0; i<bugs.size(); i++)
     bugs[i]->draw(canvas);
 }
