+++ models/Stupid_Project/stupid09.cc
@@ -107,7 +107,7 @@ void StupidModel::addBugs(TCL_args args)
 void StupidModel::moveBugs()
 {
   random_shuffle(bugs.begin(),bugs.end());
-  for (int i=0; i<bugs.size(); i++)
+  for (size_t i=0; i<bugs.size(); i++)
     bugs[i]->move();
   tstep++;
 }
@@ -143,7 +143,7 @@ void StupidModel::draw(TCL_args args)
   eco_string canvas=args;
   tclcmd c;
   c << canvas << "delete bugs\n";
-  for (int i=0; i<bugs.size(); i++)
+  for (size_t i=0; i<bugs.size(); i++)
     bugs[i]->draw(canvas);
 }
