+++ models/Stupid_Project/stupid10.cc
@@ -116,7 +116,7 @@ struct BugMore
 void StupidModel::moveBugs()
 {
   sort(bugs.begin(),bugs.end(),BugMore());
-  for (int i=0; i<bugs.size(); i++)
+  for (size_t i=0; i<bugs.size(); i++)
     bugs[i]->move();
   tstep++;
 }
@@ -152,7 +152,7 @@ void StupidModel::draw(TCL_args args)
   eco_string canvas=args;
   tclcmd c;
   c << canvas << "delete bugs\n";
-  for (int i=0; i<bugs.size(); i++)
+  for (size_t i=0; i<bugs.size(); i++)
     bugs[i]->draw(canvas);
 }
